Expand lazy-context e2e fixture coverage across all branches
