+++ core/src/test/resources/test-cases/core/e2e/restructure/11-lazy-initializer-contexts/config.yml
