+++ core/src/test/resources/test-cases/core/e2e/restructure/11-lazy-initializer-contexts/expected/LazyLambdaContextSample.java
@@ -1,0 +1,22 @@
+package io.github.lemon_ant.jharmonizer.core.e2e;
+
+public class LazyLambdaContextSample {
+    // Alphabetical order in this fixture: aDependent, bIndependent, zProvider.
+    // If tool keeps zProvider before aDependent, it means an extra dependency was added.
+    static Runnable aDependent = () -> {
+        if (LazyLambdaContextSample.zProvider != 7) {
+            throw new IllegalStateException("Unexpected zProvider value in lambda: "
+                    + LazyLambdaContextSample.zProvider);
+        }
+    };
+    static int bIndependent = 3;
+    static int zProvider = 7;
+
+    public static void main(String[] args) {
+        aDependent.run();
+        if (bIndependent != 3 || zProvider != 7) {
+            throw new IllegalStateException(
+                    "Unexpected field values: bIndependent=" + bIndependent + ", zProvider=" + zProvider);
+        }
+    }
+}
