+++ core/src/test/resources/test-cases/core/e2e/restructure/11-lazy-initializer-contexts/input/LazyLambdaContextSample.java
@@ -1,16 +1,17 @@
 package io.github.lemon_ant.jharmonizer.core.e2e;
 
-public class LazyLambdaExplicitTypeForwardReferenceSample {
-    // Alphabetical order in this fixture: aDependent, bIndependent, zProvider.
-    // If tool keeps zProvider before aDependent, it means an extra dependency was added.
+public class LazyLambdaContextSample {
+    // Intentionally placed before aDependent in input to provoke a potential false declaration dependency.
+    static int zProvider = 7;
+    // Alphabetically first field. Expected to stay first after restructuring,
+    // because explicit type-qualified access inside lambda should not force provider-before-dependent ordering.
     static Runnable aDependent = () -> {
-        if (LazyLambdaExplicitTypeForwardReferenceSample.zProvider != 7) {
+        if (LazyLambdaContextSample.zProvider != 7) {
             throw new IllegalStateException("Unexpected zProvider value in lambda: "
-                    + LazyLambdaExplicitTypeForwardReferenceSample.zProvider);
+                    + LazyLambdaContextSample.zProvider);
         }
     };
     static int bIndependent = 3;
-    static int zProvider = 7;
 
     public static void main(String[] args) {
         aDependent.run();
